+++ package.json
@@ -18,13 +18,16 @@
     "docker:down": "docker-compose -p next-template-mysql down",
     "prisma:studio": "prisma studio",
     "prisma:migration": "npx prisma format && npx prisma migrate dev && npx prisma generate",
-    "prisma:deploy": "npx prisma migrate deploy",
-    "chromatic": "npx chromatic --project-token=chpt_f10a816045e0115"
+    "prisma:deploy": "npx prisma migrate deploy"
   },
   "dependencies": {
     "@google-cloud/storage": "^6.11.0",
     "@hookform/resolvers": "^3.1.0",
+    "@mapbox/rehype-prism": "^0.8.0",
+    "@mdx-js/loader": "^2.3.0",
+    "@mdx-js/react": "^2.3.0",
     "@next-auth/prisma-adapter": "^1.0.7",
+    "@next/mdx": "^13.4.12",
     "@prisma/client": "^5.0.0",
     "@radix-ui/react-avatar": "^1.0.3",
     "@radix-ui/react-dialog": "^1.0.4",
@@ -35,6 +38,7 @@
     "@radix-ui/react-toast": "^1.1.4",
     "@storybook/addons": "^7.1.1",
     "@t3-oss/env-nextjs": "^0.4.0",
+    "@types/mdx": "^2.0.5",
     "@types/node": "20.2.5",
     "@types/react": "18.2.7",
     "@types/react-dom": "18.2.4",
@@ -52,6 +56,12 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.44.2",
+    "rehype-katex": "^6.0.3",
+    "rehype-slug": "^5.1.0",
+    "remark-breaks": "^3.0.3",
+    "remark-gfm": "^3.0.1",
+    "remark-math": "^5.1.1",
+    "remark-toc": "^8.0.1",
     "retry-request": "^5.0.2",
     "sharp": "^0.32.1",
     "tailwind-merge": "^1.14.0",
@@ -73,6 +83,7 @@
     "@storybook/react": "^7.1.1",
     "@storybook/testing-library": "^0.2.0",
     "@tailwindcss/line-clamp": "^0.4.4",
+    "@tailwindcss/typography": "^0.5.9",
     "@trivago/prettier-plugin-sort-imports": "4.1.1",
     "@types/deep-equal": "^1.0.1",
     "@types/jest": "^29.5.1",
