+++ src/app/(info)/info/page.mdx
@@ -1,0 +1,9 @@
+# MDX
+
+## 普通の マークダウン
+
+マークダウンのリスト
+
+- リスト 1
+- リスト 2
+- リスト 3
