+++ src/app/(info)/readme.md
@@ -1,0 +1,7 @@
+# このディレクトリについて
+ドキュメントの記述をしている場所
+
+### サンプル
+- [reactjs - Using MDX with NextJS 13 returning useContext error - Stack Overflow](https://stackoverflow.com/questions/75915454/using-mdx-with-nextjs-13-returning-usecontext-error)
+- [Next.js App Router で MDX を利用](https://zenn.dev/hayato94087/articles/1411f7af3eaee4)
+- [Configuring: MDX | Next.js](https://nextjs.org/docs/pages/building-your-application/configuring/mdx)
